Moved global styles variables
