+++ diff-engine/src/Components/Graphs/Calculators/Converters/Percentages/percent.styles.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { basicDark,basicLight } from "../../../global.styles";
+import { basicDark,basicLight } from "../../../../../global.styles";
 
 export const PercentDisplayContainer = styled.section`
     width:100%;
