+++ diff-engine/src/Components/Graphs/Calculators/Physics/Leverage/LeverageDisplay/HeightArrow/height-arrow.styles.jsx
@@ -1,5 +1,5 @@
 import styled, { css } from "styled-components";
-import { backgroundColors } from "../../../../../global.styles";
+import { backgroundColors } from "../../../../../../../global.styles";
 
 const { lightDark,darkLight,dark,paper } = backgroundColors
