+++ diff-engine/src/Components/Graphs/Calculators/Standard/standard.styles.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { backgroundColors } from "../../global.styles";
+import { backgroundColors } from "../../../../global.styles";
 
 const {
     paper,
