+++ diff-engine/src/Components/Graphs/Calculators/Trig/UnitCircle/unitcircle.styles.jsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import { backgroundColors } from "../../../global.styles";
+import { backgroundColors } from "../../../../../global.styles";
 
 const {
     paper,
