+++ diff-engine/src/Components/Graphs/Display/display.component.jsx
@@ -1,6 +1,6 @@
 import { useContext,useEffect } from "react"
 import { ViewContext } from "../../Context/view.context"
-import { backgroundColors } from "../global.styles"
+import { backgroundColors } from "../../../global.styles"
 import StandarMathDisplay from "../Calculators/Standard/standard.display"
 // import FractionCalc from "../Calculators/Fractions/frac.display"
 import FractionCalc from "../Calculators/Converters/Fractions/frac.display"
